requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.60.0

# PDF processing
pymupdf>=1.26.0

# LLM providers
google-generativeai>=0.8.0
anthropic>=0.40.0

# CLI
typer>=0.12.0
rich>=13.0.0
